Final README for v2.0.0
You may see that there are a lot of README and LICENSE commits just in a single day (JAN 23, 2024). This is because I was learning and experimenting, and I've got it perfect now. And this will be the final/last update for v2.0.0 of IntergalacticPro. See you in the next version!